+++ src/dashboard/apps/consent/signals.py
@@ -17,7 +17,5 @@ def handle_new_delivery_point(sender, instance, created, **kwargs):
     if created:
         Consent.objects.create(
             delivery_point=instance,
-            id_station_itinerance=instance.id_station_itinerance,
-            station_name=instance.station_name,
             provider_assigned_id=instance.provider_assigned_id,
         )